Categories are now stored after they are downlaoded.
This closes #21. After the categories are first updated, the following
times only the version is checked and new categories are downloaded if
the the version of the category list on the server is > than the last
downloaded list. If no list of categories was ever downloaded (e.g.
first run or app data removed) and the app fails to download a list of
categories, the app show na error message.